refactor(sandbox-preview): extract PathParamInput to its own module
